+++ src/main/java/com/fran/FRAN/model/dao/RelatorioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RelatorioRepository extends JpaRepository<Relatorio, Long> {
+public interface RelatorioRepository extends JpaRepository<Relatorio, Integer> {
 }
